# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mystic Weave",
    "description": "Persistent game state backend for a text-based RPG powered by a custom GPT. The GPT is the narrator; this API is the memory. d100 roll-under resolution with domain scores and competency tiers.",
    "version": "5.3.1"
  },
  "servers": [
    {
      "url": "https://mysticweave-production.up.railway.app",
      "description": "Production (Railway)"
    }
  ],
  "paths": {
...
    "/location": {
      "post": {
        "summary": "Upsert Location",
        "description": "Create or update a location via UPSERT, and persist valid world-graph connections from the payload.",
        "operationId": "upsert_location_location_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            }
          },
          "200": {
            "description": "Location updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/scene/declare_resolution": {
      "post": {
        "tags": [
          "scene"
        ],
        "summary": "Declare Scene Resolution",
        "description": "Record a scene boundary. Generates scene_id, persists structured actions and arc-progression contributions, returns envelope status for active arcs with suggest-level guidance for soft/hard caps.",
        "operationId": "declare_scene_resolution_scene_declare_resolution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclareSceneResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeclareSceneResolutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scene/record/{session_id}/{scene_id}": {
      "get": {
        "tags": [
          "scene"
        ],
        "summary": "Get Scene Record",
        "description": "Read a single scene record. Internal/admin use; not in GPT spec.",
        "operationId": "get_scene_record_scene_record__session_id___scene_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scene/records/{session_id}": {
      "get": {
        "tags": [
          "scene"
        ],
        "summary": "List Scene Records",
        "description": "Paginated list of scene records. Internal/admin use; not in GPT spec.",
        "operationId": "list_scene_records_scene_records__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneRecordsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeclareSceneResolutionRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "scene_summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Summary"
          },
          "scene_actions": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SpellCastAction"
                },
                {
                  "$ref": "#/components/schemas/WeaponAttackAction"
                },
                {
                  "$ref": "#/components/schemas/SocialRollAction"
                },
                {
                  "$ref": "#/components/schemas/PerceptionRollAction"
                },
                {
                  "$ref": "#/components/schemas/MovementAction"
                },
                {
                  "$ref": "#/components/schemas/DefenseAction"
                },
                {
                  "$ref": "#/components/schemas/GenericRollAction"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "defense": "#/components/schemas/DefenseAction",
                  "generic_roll": "#/components/schemas/GenericRollAction",
                  "movement": "#/components/schemas/MovementAction",
                  "perception_roll": "#/components/schemas/PerceptionRollAction",
                  "social_roll": "#/components/schemas/SocialRollAction",
                  "spell_cast": "#/components/schemas/SpellCastAction",
                  "weapon_attack": "#/components/schemas/WeaponAttackAction"
                }
              }
            },
            "type": "array",
            "maxItems": 20,
            "title": "Scene Actions",
            "description": "Structured actions taken during the scene"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id",
            "description": "Where the scene resolved; if omitted, server reads world.location at time of resolution."
          },
          "arc_progressed_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Arc Progressed Ids",
            "description": "Active arc IDs this scene contributed to. Server validates each exists."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "DeclareSceneResolutionRequest",
        "description": "Request payload for /scene/declare_resolution."
      },
      "DeclareSceneResolutionResponse": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "resolved_at": {
            "type": "string",
            "title": "Resolved At",
            "description": "ISO timestamp"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "turn_at_resolution": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn At Resolution"
          },
          "arc_envelope_status": {
            "items": {
              "$ref": "#/components/schemas/ArcEnvelopeStatus"
            },
            "type": "array",
            "title": "Arc Envelope Status",
            "description": "Status of all active arcs after this scene's contribution"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions",
            "description": "Suggest-level guidance: e.g., 'arc_X soft cap approaching, consider settling'"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scene_id",
          "session_id",
          "resolved_at",
          "location_id",
          "turn_at_resolution",
          "arc_envelope_status"
        ],
        "title": "DeclareSceneResolutionResponse",
        "description": "Response from /scene/declare_resolution."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "unknown"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "connections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Connections"
          },
          "threat_level": {
            "type": "integer",
            "title": "Threat Level",
            "default": 0
          },
          "known_npcs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Npcs"
          },
          "discovered": {
            "type": "boolean",
            "title": "Discovered",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "LocationData",
        "description": "Shape of the JSONB data column in the locations table."
      },
      "LocationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data": {
            "$ref": "#/components/schemas/LocationData"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "data"
        ],
        "title": "LocationResponse"
      },
...
      "MovementAction": {
        "properties": {
          "type": {
            "type": "string",
            "const": "movement",
            "title": "Type",
            "default": "movement"
          },
          "application": {
            "type": "string",
            "title": "Application",
            "description": "Canonical mobility application index (e.g., 'evasion', 'parkour')"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failure"
            ],
            "title": "Outcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "application",
          "outcome"
        ],
        "title": "MovementAction",
        "description": "A character moved with notable conditions. Maps to mobility/stealth tag."
      },
...
      "PerceptionRollAction": {
        "properties": {
          "type": {
            "type": "string",
            "const": "perception_roll",
            "title": "Type",
            "default": "perception_roll"
          },
          "application": {
            "type": "string",
            "title": "Application",
            "description": "Canonical perception application index (e.g., 'spoor_reading')"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failure"
            ],
            "title": "Outcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "application",
          "outcome"
        ],
        "title": "PerceptionRollAction",
        "description": "A character made a perception/awareness roll."
      },
...
      "SceneRecord": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "resolved_at": {
            "type": "string",
            "title": "Resolved At"
          },
          "scene_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Summary"
          },
          "scene_actions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Scene Actions",
            "description": "Raw scene actions as recorded"
          },
          "tag_advance_committed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Advance Committed"
          },
          "arc_progressed_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Arc Progressed Ids"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "turn_at_resolution": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn At Resolution"
          },
          "time_at_resolution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time At Resolution"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scene_id",
          "session_id",
          "resolved_at",
          "scene_summary",
          "scene_actions",
          "tag_advance_committed",
          "arc_progressed_ids",
          "location_id",
          "turn_at_resolution",
          "time_at_resolution"
        ],
        "title": "SceneRecord",
        "description": "Full scene record returned by GET endpoints."
      },
      "SceneRecordsListResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/SceneRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor for the next page; None if no more."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "records",
          "has_more"
        ],
        "title": "SceneRecordsListResponse",
        "description": "Paginated list of scene records for a session."
      },
...
      "SocialRollAction": {
        "properties": {
          "type": {
            "type": "string",
            "const": "social_roll",
            "title": "Type",
            "default": "social_roll"
          },
          "application": {
            "type": "string",
            "title": "Application",
            "description": "Canonical social application index (e.g., 'persuasion', 'command')"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failure"
            ],
            "title": "Outcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "application",
          "outcome"
        ],
        "title": "SocialRollAction",
        "description": "A character made a social roll. Maps to specific social tag (explicit)."
      },
      "SpellCastAction": {
        "properties": {
          "type": {
            "type": "string",
            "const": "spell_cast",
            "title": "Type",
            "default": "spell_cast"
          },
          "spell": {
            "type": "string",
            "title": "Spell",
            "description": "Canonical spell index (e.g., 'seedwake')"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failure"
            ],
            "title": "Outcome",
            "description": "Roll outcome; influences fit strength on failure"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spell",
          "outcome"
        ],
        "title": "SpellCastAction",
        "description": "A character cast a spell. Maps to spell tag (explicit) and field (implicit)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeaponAttackAction": {
        "properties": {
          "type": {
            "type": "string",
            "const": "weapon_attack",
            "title": "Type",
            "default": "weapon_attack"
          },
          "weapon": {
            "type": "string",
            "title": "Weapon",
            "description": "Canonical weapon application index (e.g., 'longbow')"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failure"
            ],
            "title": "Outcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "weapon",
          "outcome"
        ],
        "title": "WeaponAttackAction",
        "description": "A character attacked with a weapon. Maps to weapon tag (explicit) and combat group (implicit)."
      },
...
    }
  }
}